+++ src/components/common/ElementSection.scss
